QWidget: document that a widget might only get a mouseDoubleClickEvent
If two widgets are on top of each other, and the top widget disappears
in response to a press or release that is followed by a double click,
then the widget at the bottom only receives the double click event.

This is a sequence of events that the application developer that choses
to build such a UI has to take care of.

Change-Id: I440efd2cac01631de8995abf9a9fb76815de8f9a
Fixes: QTBUG-61173
Reviewed-by: Paul Wicking <paul.wicking@qt.io>