Doc: Prevent QDoc from auto-linking certain words
'macOS' appears in the documentation frequently, and because it
resembles a variable/property name, QDoc links each occurrence
of it.

'WebChannel' appears as part of a module name, but it's also
a QML type - auto-linking each word to QML documentation is
confusing.

Likewise, there's no need to link 'OpenGL' each time.

Explicit links such as \l [QML] WebChannel continue to work.

Task-number: QTBUG-79135
Change-Id: I76cc84b0076255e260aa88c244102702a48f35a6
Reviewed-by: Martin Smith <martin.smith@qt.io>