Prospective fix for Boot2Qt build in the CI
One of the parameters that Coin passes to configure is this:

    -device-option DISTRO_OPTS="hard-float boot2qt"

The configure script would "swallow" the quotes and end up calling qmake
with -device-option DISTRO_OPTS=hard-float boot2qt, causing qmake to
complain that the standalone "boot2qt" option is unknown.

Fix this by preserving the quotes in the forwarding helper functions.

Change-Id: I16098bd35acb579ebf183b0f7746af8758269e7c
Reviewed-by: Paolo Angelelli <paolo.angelelli@qt.io>