Turn a local QHash into a QMultiHash
972a0402be4d7aed0ebbd0ff56b5018cda01599d was merged before one of its
parent patches which changed this QHash into a QMultiHash. So in 5.14
it would overwrite values instead of adding new entries. No idea if it
has caused or will cause any regressions. But it was an accident so
let's fix it.

Amends 972a0402be4d7aed0ebbd0ff56b5018cda01599d.

Change-Id: I6623b0b7924024df148d5c83bcbb612f3e595f56
Reviewed-by: Lars Knoll <lars.knoll@qt.io>